db_service/requirements.txt
aiohttp
aiohttp-cors
asyncpg
python-dotenv